Changed geometry, topology, and color data type
I noticed after class 02/27/2023 the use of the triangles primitive and noticed that it was much easier to grasp as opposed to what I was using, the triangle-strips. I also noticed that I was creating a 2-unit cube, with its geometry stretching from (-1,-1,-1) to (1,1,1) giving each length 2 units. Since this assignment called for a 1-unit cube I changed the geometry and then changed the topography to match the simpler primitive type of triangles, making sure to match up the correct "facedness". I also changed the aColor variable data type when creating it with the attribute helper function to float, so that I could experiment more with the colors, making a few adjustments, though still only using 1.0 or 0.0 for each value of (r,g,b,a).